Data processing function for panel data (long format)

Data processing function for panel data (long format)

did_data_panelL(outcome, treatment, post_treatment, id_subject, id_time,
  Xcov, x_formula = NULL)

Arguments

outcome
A vector of outcome variable.

treatment
A vector of treatment indicator variable.

post_treatment
A vector of numeric time index for post treatment periods.

id_subject
A vector of subject index variable.

id_time
A vector of time index variable.

Xcov
A matrix of covariates.

Value

A list. Each element corresponds to each post-treatment period. Each element of the returned list is also a list consists of the following:

Y: outcome vector.

D: treatment vector.

formula: A list of formula for lm.

pdata: A data frame appended transformed outcomes (e.g., yd0, yd1, etc).